refactor: embed `Column` into `Relation`.
