+++ lib/postgrex_wal/messages/relation.ex
@@ -7,8 +7,59 @@ defmodule PostgrexWal.Messages.Relation do
   and first consult that cache to see if the type OID is defined there. If not, look up the type OID locally.
   """
 
+  defmodule Column do
+    @moduledoc """
+    columns message.
+
+    Int8
+    Flags for the column. Currently can be either 0 for no flags or 1 which marks the column as part of the key.
+
+    String
+    Name of the column.
+
+    Int32 (Oid)
+    OID of the column's data type.
+
+    Int32
+    Type modifier of the column (atttypmod).
+    """
+
+    alias PostgrexWal.MessageUtil
+    use TypedStruct
+
+    typedstruct enforce: true do
+      field :flags, Keyword.t()
+      field :column_name, String.t()
+      field :type_oid, integer()
+      field :type_modifier, integer()
+    end
+
+    @spec decode(columns :: binary, acc :: list()) :: [t()]
+    def decode(columns, acc \\ [])
+    def decode(<<>>, acc), do: Enum.reverse(acc)
+
+    def decode(<<flags::8, rest::binary>>, acc) do
+      [
+        column_name,
+        <<type_oid::32, type_modifier::32, rest::binary>>
+      ] = MessageUtil.binary_split(rest)
+
+      decode(
+        rest,
+        [
+          %__MODULE__{
+            flags: [{:key, flags == 1}],
+            column_name: column_name,
+            type_oid: type_oid,
+            type_modifier: type_modifier
+          }
+          | acc
+        ]
+      )
+    end
+  end
+
   use PostgrexWal.Message
-  alias PostgrexWal.Messages.Relation.Column
 
   typedstruct enforce: true do
     field :transaction_id, integer(), enforce: false
@@ -68,14 +119,14 @@ defmodule PostgrexWal.Messages.Relation do
       namespace,
       relation_name,
       <<replica_identity_setting::8, number_of_columns::16, columns::binary>>
-    ] = Util.binary_split(rest, 3)
+    ] = MessageUtil.binary_split(rest, 3)
 
     %__MODULE__{
       relation_oid: relation_oid,
-      namespace: Util.decode_namespace(namespace),
+      namespace: MessageUtil.decode_namespace(namespace),
       relation_name: relation_name,
       flags: [
-        {:setting, Util.decode_replica_identity_setting(replica_identity_setting)}
+        {:setting, MessageUtil.decode_replica_identity_setting(replica_identity_setting)}
       ],
       number_of_columns: number_of_columns,
       columns: Column.decode(columns)
